+++ src/benchmarks/ParallelReadHistoBenchmark.cpp
@@ -66,7 +66,7 @@ void ParallelReadHistoBenchmark::run() {
 
     // Verify
     uint64_t sum = 0;
-    for (uint64_t i = 0; i < n_threads * Constants::N_PARTITIONS; ++i)
+    for (uint64_t i = 0; i < static_cast<uint64_t>(n_threads) * Constants::N_PARTITIONS; ++i)
         sum += global_histogram[i];
 
     cout << "Tuples read: " << sum  << endl;
